requirements.txt
onnxruntime-gpu
opencv-python
PyQt5
shapely
tqdm
numpy
